Fixing for issues seen when building for Ubuntu.
The Ubuntu version of autotools is picky about the subdir's and throws many warnings during a reconf.
Added the right AM_INIT_AUTOMAKE([subdir-objects]) to the configure.ac file to stop it from complaining.

The version of gcc is also more picky about includes so I had to explicitly include the unistd.h for the
sleep statement in the example progam.