RHWEB-229 Eclipse & Redhawk IDEs should now be able to build library without reconf run
The autotools configure.ac and makefile.am were modified to add the m4 folder to the
location searched so the reconf script is no longer necessary but was kept for backwards
compatibility and for those used to typing reconf.  A user may now instead use autoreconf -i
in the same way the IDEs do.