CA-381 Fixing strictness of automake to fix broken build.
A previous commit (CA-312) removed the README from librfsimulators which caused the build to fail due to the strictness level of the gnu automake tooling. librfsimulators was setup as a gnu automake project and because of that was set to the default (strict) gnu automake rules. Adding "foreign" to the configure.ac file will fix this and put it more in line with how automake is configured in other REDHAWK related assets. It will also eliminate the need for duplicate or unused files like COPYING (dup of LICENSE), AUTHORS, and NEWS which are all required by strict gnu automake.